eye 컴포넌트 테스트용 업로드

+++ RoguelightSpeedRun20D/Assets/_/minwoo/01.Scripts/01.Item/01.Struct/ShoesData.cs
@@ -2,11 +2,11 @@
 public struct ShoesData
 {
     int maxHp;
-    float speed;
+    int speed;
     int maxStamina;
     float staminaRegen;
 
-    public ShoesData(int maxHp, float speed, int maxStamina, float staminaRegen)
+    public ShoesData(int maxHp, int speed, int maxStamina, float staminaRegen)
     {
         this.maxHp = maxHp;
         this.speed = speed;
@@ -15,7 +15,7 @@ public struct ShoesData
     }
 
     public int MaxHp { get => maxHp; set => maxHp = value; }
-    public float Speed { get => speed; set => speed = value; }
+    public int Speed { get => speed; set => speed = value; }
     public int MaxStamina { get => maxStamina; set => maxStamina = value; }
     public float StaminaRegen { get => staminaRegen; set => staminaRegen = value; }
 }
